// src/shared/generated/CoreAppSnapshotRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BrowserRoleStatusRecord } from "./BrowserRoleStatusRecord";
import type { BrowserRuntimeSnapshot } from "./BrowserRuntimeSnapshot";
import type { CoreStateSnapshotRecord } from "./CoreStateSnapshotRecord";
import type { MacroRunStatus } from "./MacroRunStatus";
import type { RuntimeWindowTabSnapshotRecord } from "./RuntimeWindowTabSnapshotRecord";

export type CoreAppSnapshotRecord = { revision: number, stateRevision: number, runtimeRevision: number, state: CoreStateSnapshotRecord, browserRuntime: BrowserRuntimeSnapshot, logicalWindows: Array<RuntimeWindowTabSnapshotRecord>, roleStatuses: Array<BrowserRoleStatusRecord>, macroStatuses: Array<MacroRunStatus>, };