// src/shared/generated/GameWindowTabRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GameWindowRoleSlotRecord } from "./GameWindowRoleSlotRecord";

export type GameWindowTabRecord = { id: string, tabType: "role" | "workspace", sourceId: string, name: string, roleSlots: Array<GameWindowRoleSlotRecord>, hidden: boolean, audioMuted: boolean, };