// src/shared/generated/RuntimeWindowProvisionTargetRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StatePixelBoundsRecord } from "./StatePixelBoundsRecord";

export type RuntimeWindowProvisionTargetRecord = { persistedName?: string, displayId: number, scaleFactor: number, workArea: StatePixelBoundsRecord, bounds: StatePixelBoundsRecord, presentation: "normal" | "maximized" | "fullscreen", };